ml-api/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Pillow==10.3.0
Werkzeug==3.0.2
gunicorn==22.0.0
